refactor: エラーページをコンポーネント化
