+++ frontend/src/pages/404.tsx
@@ -1,27 +1,12 @@
-import { Button, Typography } from '@mui/material'
-import { useRouter } from 'next/router'
-import { FlexLayout } from '@/components/layouts/common/FlexLayout'
+import { ErrorPage } from '@/components/pages/ErrorPage'
 
 const Custom404 = () => {
-  const router = useRouter()
-
   return (
-    <FlexLayout justifyContent="center" alignItems="center" direction="column">
-      <Typography variant="h1" marginTop={8}>
-        404
-      </Typography>
-      <Typography variant="h3">Not Found</Typography>
-      <Typography>お探しのページは見つかりません</Typography>
-      <Button
-        variant="contained"
-        color="primary"
-        onClick={() => router.back()}
-        sx={{ marginTop: 4 }}
-        aria-label="前のページに戻る"
-      >
-        前に戻る
-      </Button>
-    </FlexLayout>
+    <ErrorPage
+      code={404}
+      message="Not Found"
+      description="お探しのページは見つかりませんでした。URLを確認してください。"
+    />
   )
 }
