+++ frontend/src/pages/500.tsx
@@ -1,29 +1,12 @@
-import { Button, Typography } from '@mui/material'
-import { useRouter } from 'next/router'
-import { FlexLayout } from '@/components/layouts/common/FlexLayout'
+import { ErrorPage } from '@/components/pages/ErrorPage'
 
 const Custom500 = () => {
-  const router = useRouter()
-
   return (
-    <FlexLayout justifyContent="center" alignItems="center" direction="column">
-      <Typography variant="h1" marginTop={8}>
-        500
-      </Typography>
-      <Typography variant="h3">Internal Server Error</Typography>
-      <Typography>
-        サーバーで問題が発生しました。時間をおいて再度お試しください。
-      </Typography>
-      <Button
-        variant="contained"
-        color="primary"
-        onClick={() => router.back()}
-        sx={{ marginTop: 4 }}
-        aria-label="前のページに戻る"
-      >
-        前に戻る
-      </Button>
-    </FlexLayout>
+    <ErrorPage
+      code={500}
+      message="Internal Server Error"
+      description="サーバーで問題が発生しました。時間をおいて再度お試しください。"
+    />
   )
 }
